Phase 5 step 4: fix Ecos function-command merge bug
CommandRouter::handleEcosFunctionCommand() overwrote the entire 32-bit
function bitmap on every Ecos event, using only that one event's
functions value - since a real Ecos event typically only reports the
single function that changed, every other already-known function would
silently reset to 0. Now takes a functions_mask parameter and merges,
only touching the bits this specific event actually reported.

Found a second, more fundamental bug in the same area while wiring this
up: functions_mask was uint8_t (only 8 bits, but functions span F0-F31)
with a plain int shift rather than the 1UL << idiom already used
correctly one line above for functions itself - so any function at F8
or above silently truncated to 0 on assignment, never registering in
the mask regardless of merge logic. Widened to uint32_t and fixed the
shift.

No prior test coverage existed for func[] parsing at all. Added 4 parser
tests (including a regression test for the truncation bug at F31) and 2
CommandRouter tests exercising the actual merge. Native suite 118/118.

Firmware builds clean and is flashed, but still needs a live test: turn
two different functions on via Ecos one at a time and confirm both stay
on.

+++ xpressnet_ecos_bridge/protocols/ecos/ecos_message_parser.h
@@ -51,8 +51,14 @@ struct EcosReply {
     uint8_t direction;              // 0=reverse, 1=forward (invalid if not set)
     uint32_t functions;             // Bitmap F0-F31
 
-    uint8_t functions_mask;         // Which function bits were actually set in this reply
-                                    // (so we don't overwrite unset functions when backfilling)
+    // Which function bits were actually reported in this reply (so a caller
+    // can merge instead of overwriting unset functions when backfilling).
+    // Real bug found in codebase audit 2026-08-03: this was uint8_t, which
+    // can only represent F0-F7 - `1 << fn_index` for fn_index 8-31 silently
+    // truncated to 0 on assignment, so any reported bit at F8 or above
+    // never actually registered in the mask. Must match `functions`'s own
+    // 32-bit width.
+    uint32_t functions_mask;
 
     SystemStatus system_status;     // Valid only if has_system_status
